Fixed compiling issue
Licor 6800s seem to output different number of columns depending on the settings. While these settings may not affect the data, they do affect compilation. Now this function will compile files with different numbers of columns. Please note that if a file is missing a column, it will fill in NAs for that file.